perf: only subscribe to policies in IOURequestStepConfirmation while the participant picker is open
